final changes

+++ app/Http/Controllers/PlaylistController.php
@@ -82,20 +82,30 @@ class PlaylistController extends Controller {
         $pl = Playlist::where('id', $id)->first();
         if (empty($pl))         { return response(null, Response::HTTP_NOT_FOUND); }
 
-        
+        $queue = EncodeQueue::where('playlistId', $id)->get();
+        if ($queue->count() > 0)      { return response(null, Response::HTTP_CONFLICT); }
+
         // Borrar contenidos de la lista
         $media = Upload::where('playlistId', $id)->get();
 
-        foreach ($media as $k => $v) {
+        foreach ($media as $v) {
             $file = PATH_MEDIA . $v->filename;
             if (file_exists($file))  { unlink($file); }
 
             $thumb = PATH_THUMBS . pathinfo($v->filename, PATHINFO_FILENAME) . '.webp';
             if (file_exists($thumb)) { unlink($thumb); }
-
-            Upload::where('id', $v->id)->delete();
         }
+        Upload::where('playlistId', $id)->delete();
 
+        // Borrar cola de codificación
+
+        foreach ($queue as $v) {
+            $file = PATH_QUEUE . $v->filename;
+            if (file_exists($file))  { unlink($file); }
+        }
+        EncodeQueue::where('playlistId', $id)->delete();
+
+        // Borrar lista
         Playlist::where('id', $id)->delete();
         
         return response()->json(['playlist' => $pl->name, 'media' => $media], Response::HTTP_OK);
